onestl: in class BitSet, pack bits in uint8_t instead of size_t onestl: rewrite BitSet::set() and Graph::set()

+++ include/onestl/bitset.hpp
@@ -36,7 +36,7 @@ namespace onestl {
 class BitSet {
 
 protected:
-  typedef size_t T;
+  typedef uint8_t T;
 
   enum : size_t {
     sizeofT = sizeof(T),
